beta ver 1.3

+++ media/JS/utility/Firebase.js
@@ -34,6 +34,8 @@ export default class Firebase {
         return _currentTableId;
     }
     static initialize() {
+        if(!Firebase.haveInternet()) return;
+
         firebase.initializeApp({
             apiKey: 'AIzaSyA8X5BGIHdic4rowTSnVhx_OrdHSkonngQ',
             authDomain: 'chess-club-online.firebaseapp.com',
@@ -44,6 +46,8 @@ export default class Firebase {
         _databaseCollectionTables = Firebase.db.collection(`tables`);
     }
     static setUser(userInfo = {}, successCompletedCallback = (id) => { }, failCompletedCallback = (error) => { }) {
+        if(!Firebase.haveInternet()) return;
+        
         let customUser = new User(userInfo.name, userInfo.email, userInfo.password);
         let newUser = Firebase.convertCustomObjToGenericObj(customUser)
         let p = Firebase.dbUsers.add(newUser);
@@ -56,6 +60,8 @@ export default class Firebase {
     }
     // for sign in feature
     static authenticateUser(givenUserName = `phong`, givenPassword = `12345`, completedCallback = (isRight, id, userData) => { }) {
+        if(!Firebase.haveInternet()) return;
+        
         let isGivenPasswordRight = false;
         let p = Firebase.dbUsers.where(`name`, `==`, givenUserName).get();
         p.then((querySnapshot) => {
@@ -73,6 +79,8 @@ export default class Firebase {
     }
     // for sign up feature
     static findNameAndEmailDuplicate(givenName = "phong", givenEmail = "fun@mail.com", givenPassword = `12345`, completedCallback = (isNR, isER, uInfo) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let isNameDuplicate = false;
         let isEmailDuplicate = false;
         let userInfo = new User(givenName, givenEmail, givenPassword);
@@ -91,6 +99,8 @@ export default class Firebase {
         });
     }
     static queryAllTable(completedCallback = (allTables) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let p = Firebase.dbTalbes.get();
         p.then((querySnapshot) => {
             let documents = querySnapshot.docs;
@@ -98,6 +108,8 @@ export default class Firebase {
         });
     }
     static onSnapshotWithId(id = `id`, changedCallback = (tableData) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         _unSubcribeSnapshot = Firebase.dbTalbes.doc(id)
             .onSnapshot((doc) => {
                 if (doc.exists) {
@@ -108,6 +120,8 @@ export default class Firebase {
     }
 
     static updateTableProperty(tableId, propertyObj, resolveCallback = () => { }, failCallback = (errorCode) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let ref = Firebase.dbTalbes.doc(tableId);
         ref.update(propertyObj)
             .then(() => {
@@ -119,6 +133,8 @@ export default class Firebase {
     }
 
     static updateMove(tableId, pieceLastMove, pieceMove, resolveCallback = () => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let obj = {};
         let move = "ownerMove";
         let lastMove = "ownerLastMove";
@@ -145,6 +161,8 @@ export default class Firebase {
     }
 
     static updateCurrentUserData(userId, propObj, resolveCallback = () => { }, failCallback = (errorCode) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let ref = Firebase.dbTalbes.doc(userId);
         ref.update(propObj)
             .then(() => {
@@ -156,6 +174,8 @@ export default class Firebase {
     }
 
     static setCurrentUserData(userId, userObj, resolveCallback = (docData) => { }, failCallback = (errorCode) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let p = Firebase.dbUsers.doc(userId).set(userObj);
         p.then(() => {
             resolveCallback();
@@ -166,6 +186,8 @@ export default class Firebase {
     }
 
     static getTable(tableId, resolveCallback = (docData) => { }, failCallback = (errorCode) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let p = Firebase.dbTalbes.doc(tableId).get();
         p.then((doc) => {
             if (doc.exists) {
@@ -178,6 +200,8 @@ export default class Firebase {
             });
     }
     static setTable(tableId, tableObj, resolveCallback = () => { }, failCallback = (error) => { }) {
+        if(!Firebase.haveInternet()) return;
+
         let p = Firebase.dbTalbes.doc(tableId).set(tableObj);
         p.then(() => {
             resolveCallback();
@@ -187,7 +211,9 @@ export default class Firebase {
             });
     }
     static deleteTable(tableId, resolveCallback = () => { }, failCallback = (errorCode) => { }) {
-        Firebase.updateTableProperty(tableId, { "tableId": -1 }, () => {
+        if(!Firebase.haveInternet()) return;
+
+        Firebase.updateTableProperty(tableId, { "tableId": -1, }, () => {
             // actually delete code below here
             Firebase.dbTalbes.doc(tableId).delete()
                 .then(() => {
@@ -215,5 +241,10 @@ export default class Firebase {
             genericObj[property] = customObj[property];
         }
         return genericObj;
+    }
+    static haveInternet() {
+
+
+        return true;
     }
 }
